requirements.txt
flask==2.3.3
requests==2.31.0
numpy>=1.21.0
Pillow>=9.0.0
pytest>=7.0.0
apsi
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.1.0
torch
pandas